Adds ergonomic wrapper for Context
This adds an ergonomic wrapper for Context, which uses an OsStack type
stack. I was unable to make it generic across all stack types, but this
does provide some nice usage. It also supports terminating and prevents
the fiber/thread from being run once it has terminated, and returns a
value indicating if it has terminated.